third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-paint/timeline-paint-with-layout-invalidations-on-deleted-node-expected.txt

Tests the Timeline API instrumentation of layout invalidations on a deleted node.



Running: testLocalFrame
paint invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: http://127.0.0.1:8000/devtools/tracing/timeline-paint/timeline-paint-with-layout-invalidations-on-deleted-node.html:9}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "BODY"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: http://127.0.0.1:8000/devtools/tracing/timeline-paint/timeline-paint-with-layout-invalidations-on-deleted-node.html:11}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Style changed, stackTrace: http://127.0.0.1:8000/devtools/tracing/timeline-paint/timeline-paint-with-layout-invalidations-on-deleted-node.html:12}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
    {
        cause : {reason: Removed from layout, stackTrace: http://127.0.0.1:8000/devtools/tracing/timeline-paint/timeline-paint-with-layout-invalidations-on-deleted-node.html:13}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "#text"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
    {
        cause : {reason: Removed from layout, stackTrace: http://127.0.0.1:8000/devtools/tracing/timeline-paint/timeline-paint-with-layout-invalidations-on-deleted-node.html:13}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]

Running: testSubframe
second paint invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: http://127.0.0.1:8000/devtools/tracing/timeline-paint/timeline-paint-with-layout-invalidations-on-deleted-node.html:20}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Style changed, stackTrace: http://127.0.0.1:8000/devtools/tracing/timeline-paint/timeline-paint-with-layout-invalidations-on-deleted-node.html:21}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
    {
        cause : {reason: Removed from layout, stackTrace: http://127.0.0.1:8000/devtools/tracing/timeline-paint/timeline-paint-with-layout-invalidations-on-deleted-node.html:22}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]
